feat: rules tests and look up if user is bot
Add function to run each user through a battery of tests to generate a factor that represents how likely they are to be a bot.

Add another function to look up if a user is a bot.

Use run_tests() to run the tests, then use get_test_results() to get the list of results and is_bot(id) to look up if a user is a bot.