fix: function overload
The get_test_results() function was overloaded which is not a feature in Python. get_rest_results() has been removed and set_test_results() has been renamed to test_results() and returns a list of dict. set_bot_results() is also renamed to bot_results() and returns a list of IDs of users who are confirmed bots.